+++ app/layout.tsx
@@ -7,8 +7,8 @@ import LoginEventHandler from '@/components/LoginEventHandler';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Weardrobe',
+  description: '의류 특화 중고거래 플랫폼',
 };
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
